training/requirements.txt
absl-py==0.8.1
astor==0.8.0
boto3==1.9.249
botocore==1.12.249
docutils==0.15.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
jmespath==0.9.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.2
opt-einsum==3.1.0
Pillow==6.2.0
protobuf==3.10.0
python-dateutil==2.8.0
PyYAML==5.1.2
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
tensorboard==1.15.0
tensorflow==1.15.0rc3
tensorflow-estimator==1.15.1
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
